--- reliable_tcp_data_streams/ping.go
@@ -1,50 +1,0 @@
-package ch03
-
-import (
-	"context"
-	"io"
-	"time"
-)
-
-const defaultPingInterval = 30 * time.Second
-
-func Pinger(ctx context.Context, w io.Writer, reset <-chan time.Duration) {
-	var interval time.Duration
-	select {
-	case <-ctx.Done():
-		return
-	case interval = <-reset: // pulled initial interval off res
-	default:
-	}
-	if interval <= 0 {
-		interval = defaultPingInterval
-	}
-
-	timer := time.NewTimer(interval)
-	defer func() {
-		if !timer.Stop() {
-			<-timer.C
-		}
-	}()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case newInterval := <-reset:
-			if !timer.Stop() {
-				<-timer.C
-			}
-			if newInterval > 0 {
-				interval = newInterval
-			}
-		case <-timer.C:
-			if _, err := w.Write([]byte("ping")); err != nil {
-				// track and act on consecutive timeouts here
-				return
-			}
-		}
-
-		_ = timer.Reset(interval)
-	}
-}
